+++ src/components/ChannelList/hooks/useChannelUpdatedListener.ts
@@ -50,6 +50,5 @@ export const useChannelUpdatedListener = <
     return () => {
       client.off('channel.updated', handleEvent);
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [customHandler]);
+  }, [client, customHandler, forceUpdate, setChannels]);
 };